// src/types/store.ts
export interface TPartyUserInfoType {
  id: string;
  email?: string;
  nickName?: string;
  profileImage?: string;
  isPrivateEmail?: boolean;
}

export interface AuthInfoType {
  thirdPartyId?: TPartyUserInfoType;
  username?: string;
  email?: string;
  origin?: string;
  handleSubmit?: any;
}
// ...
export interface MyCenter {
  id: string;
  address: string;
  name: string;
}

export interface MyCenterItem {
  id: string;
  createdAt: string;
  center: MyCenter;
}
// ...
export interface Hearts {
  count: number;
  used: boolean;
  [keyword: string]: HeartClassData | number | boolean;
}

export interface Bookmarks {
  count: number;
  used: boolean;
  [keyword: string]: boolean | number;
}

export interface PostSavedObj {
  count: number;
  used: boolean;
  [keyword: string]: string | number | boolean;
}
export interface HeartClassData {
  id: string;
  title: string;
  tagSearchable: string | null;
  regionSearchable: string;
  host: {
    id: string;
  };
  dateStart: number;
  timeStart: number;
  classFee: number;
}

export interface NewsData {
  id: string;
  newsType: string;
  createdAt: string;
  newsStatus: string;
  newsTitle?: string;
  newsURL?: string;
  newsScreen?: string;
  newsBanner?: {
    bucket: string;
    region: string;
    key: string;
  };
  validDate: string;
  extraInfo?: string;
}

export interface IsFirstObject {
  [screenName: string]: boolean;
}

export interface NotiClassData {
  items: NewClassNotiType[];
  lastTime: number;
  // count: number;
}

export interface NotiPostData {
  items: NewPostNotiType[];
  lastTime: number;
  // count: number;
}
export interface AuthStoreType {
  authState: string;
  user: CognitoUser | any | null;
  profileName: string | null;
  identityId: string | null;
  profileUpdated: boolean;
  error?: string;
  message?: string;
  authInfo: AuthInfoType | null;
  bookmark: Bookmarks | null;
  simpleSearchHistory: SearchHistory[] | null;
  detailSearchHistory: SearchHistory[] | null;
  appState: AppStateStatus;
  subscribedTags: string[] | null;
  newClassNotis: NotiClassData | null;
  newPostNotis: NotiPostData | null;
  newGenNotisAvailable: boolean;
  newCommNotisAvailable: boolean;
  newClassNotisAvailable: boolean;
  newPostNotisAvailable: boolean;
  newChats: string[];
  oneSignalTags: { [key: string]: string } | null;
  lastReadClassAt: number | null;
  lastSubscribedTagsUpdated: number | null;
  hearts: Hearts | null;
  heartsUpdated: boolean;
  isFirst: IsFirstObject | {};
  isInternetReachable: boolean;
  isOnBoardingFinished: boolean;
  isNewButtonTouched: boolean;
  readyToServe: boolean;
  postLikes: PostSavedObj | null;
  postBookmark: PostSavedObj | null;
  appearance: AppearanceType | ColorSchemeName;
}

export enum AppearanceType {
  LIGHT = 'light',
  DARK = 'dark',
}
export interface NewClassNotiType {
  title: string;
  classId: string;
  hostId: string;
  keywordsMatched: string[] | null;
  region: string;
  isRegionMatched: boolean;
  createdAt: string;
  createdAtEpoch: number;
}
export interface NewPostNotiType {
  postTitle: string;
  postId: string;
  postAuthorId: string;
  keywordsMatched: string[] | null;
  createdAt: string;
  createdAtEpoch: number;
}
// ...
export interface StoreAction {
  type: string;
  [key: string]: any;
}